add splice hotspots to mutation table and oncoprint
