+++ packages/react-mutation-mapper/src/component/column/Annotation.tsx
@@ -9,7 +9,7 @@ import {
     IMyCancerGenomeData,
     IOncoKbData,
     is3dHotspot,
-    isRecurrentHotspot,
+    isNon3dHotspot,
     MobxCache,
     Mutation,
     RemoteData,
@@ -170,7 +170,7 @@ export function getAnnotationData(
                 hotspotData &&
                 hotspotData.result &&
                 hotspotData.status === 'complete'
-                    ? isRecurrentHotspot(mutation, hotspotData.result)
+                    ? isNon3dHotspot(mutation, hotspotData.result)
                     : false,
             is3dHotspot:
                 hotspotData &&
